Extract a `render` module out of `mvt`
Two thirds of `mvt.rs` was not about Mapbox Vector Tiles. Drawing a
geometry the way the style says to is the same job whether the geometry
arrived in a vector tile, a GeoJSON file or a KML one, and `geojson.rs`
already imported `render_line` and friends out of a module named after a
format it does not use.

`resolve_text` showed the seam most clearly. It belongs with `Text` and
`OccupiedAreas` in `text.rs`, but it had to live in `mvt.rs` to reach
`ShapeOrText`, which is to say a generic function was dragged into the MVT
module by a dependency pointing the wrong way. `draw_text` lands in
`text.rs` now, and `resolve_text` sits next to the type it consumes.

What is left in `mvt.rs` decodes tiles: reading features out of a layer,
turning MVT property values into JSON ones, and the extent that tile
coordinates are expressed in. Its `Error` loses the tessellation variant,
which is the renderer's to raise, along with three variants nothing has
been constructing.

Nothing about the public API moves - the crate root re-exports the same
names from the new module - so `walkers_extras` compiles untouched, which
is the evidence that this was only a move.
